Heroku

+++ app/Http/Controllers/PagesController.php
@@ -3,18 +3,21 @@
 namespace App\Http\Controllers;
 use Auth;
 use App\Blog;
+use App\User;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
 
 class PagesController extends Controller
 {
     //
-    public function redirect(){ 
-        if(Auth::guest()){
-            return redirect('/login');
-        }else{
-            return redirect('/home');
-        }
+    public function redirect(){
+        // if(Auth::guest()){
+        //     return redirect('/login');
+        // }else{
+        //     return redirect('/home');
+        // }
+        return redirect('/home');
     }
 
     public function create(){ 
@@ -26,7 +29,10 @@ class PagesController extends Controller
         
     }
 
-    public function store(Request $request){ 
+    public function store(Request $request){
+        if (Auth::guest()) {
+            return redirect('/login');
+        } 
         $this->validate($request,[
             'title'=>'required',
             'subtitle'=>'required',
@@ -36,6 +42,8 @@ class PagesController extends Controller
         $blog->title=$request->input('title');
         $blog->subtitle=$request->input('subtitle');
         $blog->body=$request->input('body');
+        $blog->user_id = Auth::id();
+        $blog->user_name = Auth::user()->name;
         $blog->save();
         return redirect('/home')->with('success','Blog Created');
         
@@ -43,6 +51,10 @@ class PagesController extends Controller
 
     public function delete(Request $request,$id){
         $blog = Blog::find($id);
+        if(Auth::id() != $blog->user_id && !Auth::user()->is_admin){
+            abort(403);
+        }
+
         $blog->delete();
         return redirect('/home')->with('success','Blog Removed');
     }
@@ -54,11 +66,22 @@ class PagesController extends Controller
             'body'=>'required',
         ]);
         $blog = Blog::find($id);
+        if (Auth::id() != $blog->user_id && !Auth::user()->is_admin) {
+            abort(403);
+        }
         $blog->title=$request->input('title');
         $blog->subtitle=$request->input('subtitle');
         $blog->body=$request->input('body');
         $blog->save();
         return redirect('/show/'.$id)->with('success','Blog Updated');
+        
+    }
+
+    public function user(Request $request, $id)
+    {
+        $user = User::find($id)->name;
+        $blogs = Blog::where('user_id', $id)->orderBy('created_at', 'DESC')->paginate(2);
+        return view('pages.user')->with('blogs', $blogs)->with('user',$user);
         
     }
 }
